requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
langgraph>=0.2,<2.0
langgraph-checkpoint-redis>=0.0.6,<1.0
langchain-openai>=0.2,<2.0
langchain-community>=0.3,<1.0
langchain-google-community>=2.0,<3.0
langchain-text-splitters>=0.3,<1.0
python-dotenv>=1.0,<2.0
openai>=1.50,<3.0
pydantic>=2.8,<3.0
pydantic-settings>=2.4,<3.0
duckduckgo-search>=6.3,<9.0
tavily-python>=0.5,<1.0
requests>=2.32,<3.0
faiss-cpu>=1.8,<2.0
numpy>=1.26,<3.0
pandas>=2.2,<3.0
aiohttp>=3.10,<4.0
httpx>=0.27,<1.0
tqdm>=4.66,<5.0
more-itertools>=10.0,<11.0

# Development and CI
pytest>=8.3,<9.0
ruff>=0.8,<1.0
mypy>=1.13,<2.0
types-requests>=2.32,<3.0
